redirect: setup Dockerfile
To get the build running, there were several changes that had to be
done.

The Gemfile.lock was causing issues in Linux container, so deleting it.
Several of the dependencies were missing in the Gemfile so added them.

Also created a README to show usage for container. The container is
still not usable due to missing instructions for the database setup but
will hopefully add them soon.